Revert "Merge dev in stable (#89)"
This reverts commit 0401a3d85c9dfb20ef5f557b4b7e6973bf8949b9.

+++ runtime/src/altv-c-api/altv.h
@@ -3,6 +3,7 @@
 #include "baseobject.h"
 #include "blip.h"
 #include "checkpoint.h"
+#include "entity.h"
 #include "player.h"
 #include "server.h"
 #include "vehicle.h"
@@ -13,7 +14,7 @@
 extern "C"
 {
 #endif
-EXPORT void FreeUIntArray(alt::Array<uint32_t> *array);
+EXPORT void FreeUIntArray(alt::Array<uint16_t> *array);
 EXPORT void FreePlayerPointerArray(alt::Array<alt::IPlayer*> *array);
 EXPORT void FreeVehiclePointerArray(alt::Array<alt::IVehicle*>* array);
 EXPORT void FreeStringViewArray(alt::Array<alt::StringView> *array);
